requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4

